+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Changed
 
-- Now commands show their output synchronously, giving the feel of a faster CLI.
+- Now all commands show their output synchronously, giving the feel of a faster CLI.
 
 ### Fixed